Styling, db requests
Toyed a bit with the styling of the Add-Blog-Post form.
Created abstract function for fetch requests towards the back-end.
Added functionality to upload data from the Blog-Post form to the server and to Mongo.
Added functionality to get all current blog posts and display them when the Blog section is being viewed.